moved components folder into src
